chore: fixed to modular schema for img generation endopoint
