+++ .gitignore
@@ -200,3 +200,5 @@ psd
 thumb
 sketch
 comandos.txt
+
+backend/imagenes/
